+++ useScroll/ScrollbarTrack.js
@@ -74,17 +74,23 @@ const ScrollbarTrack = forwardRef((props, ref) => {
 			if ((vertical && (isUpDown || isPageKey)) || (!vertical && (isLeftRight))) {
 				// Do nothing when (!vertical && pageKey)
 
-				if (!ev.repeat && announceRef.current.announce) {
-					announceRef.current.announce(
-						(isDown(keyCode) || isPageDown(keyCode)) && $L('DOWN') ||
-						(isUp(keyCode) || isPageUp(keyCode)) && $L('UP') ||
-						isLeft(keyCode) && $L('LEFT') ||
-						$L('RIGHT') // the case that isRight(keyCode) is true
-					);
-				}
-
 				if (ev.repeat || !isNaN(scrollProgress) && ((scrollParam.isForward && scrollProgress !== 1) || (!scrollParam.isForward && scrollProgress !== 0))) {
 					consumeEventWithScroll(scrollParam, ev);
+
+					setTimeout(()=>{
+						const updatedScrollProgress = Number(ref.current && ref.current.style.getPropertyValue('--scrollbar-thumb-progress-ratio'));
+						const horizontalReachLeftMost = rtl ? updatedScrollProgress === 1 : updatedScrollProgress === 0;
+						const horizontalReachRightMost = rtl ? updatedScrollProgress === 0 : updatedScrollProgress === 1;
+
+						if (!ev.repeat && announceRef.current.announce) {
+							announceRef.current.announce(
+								(isDown(keyCode) || isPageDown(keyCode)) && ((updatedScrollProgress === 1 && $L('DOWNMOST')) || $L('DOWN')) ||
+								(isUp(keyCode) || isPageUp(keyCode)) && ((updatedScrollProgress === 0 && $L('UPMOST')) || $L('UP')) ||
+								isLeft(keyCode) && ((horizontalReachLeftMost && $L('LEFTMOST')) || $L('LEFT')) ||
+								((horizontalReachRightMost && $L('RIGHTMOST')) || $L('RIGHT')) // the case that isRight(keyCode) is true
+							);
+						}
+					}, 500); // Wait for finishing scroll animation
 				}
 			}