+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Changed
+
+- `sandstone/Scroller` scrollbar thumb to read out "Press ok button button to read text" additionally when `focusableScrollbar` prop is `byEnter`
+- `sandstone/Scroller` scrollbar thumb to read out 'leftmost', 'rightmost', 'topmost', or 'downmost' when reaching the end of the scroll
+- `sandstone/Picker` and `sandstone/RangePicker` to read out `title`
+
 ## [2.5.0-beta.1] - 2022-05-31
 
 ### Added